feat: 입력값 검증을 위한 클래스 생성 및 기존 메서드 교체
